MAJOR UPDATE (And still not done), Backup before you update.
This update (before going to 1.0 version, is needed to clean up the internate text handlings once and for all.
Instead of dealing with all kind of text flavours (FormattedString from DrawBot) in main stotage, there is not one BabelString class that keeps the text runs and style. Contexts simply need to implement fromBabelString and asBabelString, and a limited set of methods, such as context.textSize and context.textLines.
This means that the rendering (and hyphenations) still is context dependent, but the internal storage of the result is entirely BabelString/BabelRun/BabelLine base, without binding to the context that generated text lines or determined the text size.

Sorry if this breaks your code right now, but it's the last major update we need to do, before releasing PageBot 1:0
Some deprecated classes are still here as reference code, but will be removed soon.

Note some changes:
- Element TextBox merged into Element Text, so the class does not exist anymore
- String classes, such as DrawBotString and FlatString are no longer needed.
- BabelString and Elements can store their own local context, if that is convenient in the future, but overall they will query the defined context in Document, if they need to render context specific text wrapping