Remove originTop feature from the PageBot system.
It never worked properly, creates much confusion and likely to cost a lost more time to get it right. Maybe something for version 2.0
Standard now is origin in bottom-left, as it is with DrawBot.
Exception is the behavior of Text elements, where the alignment is on the baseline of the top line of a BabelString block.

Working on positioning of text still.